+++ src/Pybind11Wraps/Plane.py
@@ -1,7 +1,8 @@
 from PYB11Generator import *
 
-class Plane3d:
-    """The PolyClipper 3D plane.
+@PYB11template("VectorType", "VA")
+class Plane:
+    """The PolyClipper plane.
 
 A plane is defined by a signed scalar distance (dist) and unit normal (normal).
 The distance d represents the signed shortest distance from the plane to the 
@@ -22,7 +23,7 @@ Planes also optionally keep an integer ID, which is used during clipping
 operations to track which plane(s) are responsible for each vertex."""
 
     PYB11typedefs = """
-typedef Vector3d Vector;
+    using Vector = %(VectorType)s;
 """
 
     #---------------------------------------------------------------------------
@@ -48,7 +49,7 @@ typedef Vector3d Vector;
         "Construct using a point in the plane (p), unit normal (nhat), plane ID (id)"
 
     def pyinit4(self,
-                rhs = "const Plane3d&"):
+                rhs = "const Plane<%(VectorType)s, %(VA)s>&"):
         "Copy constructor"
         
     #---------------------------------------------------------------------------
@@ -69,7 +70,7 @@ typedef Vector3d Vector;
     #---------------------------------------------------------------------------
     # Methods
     #---------------------------------------------------------------------------
-    @PYB11implementation('''[](const Plane3d& self) { return "{" + std::to_string(self.dist) + ", (" + std::to_string(self.normal.x) + ", " + std::to_string(self.normal.y)+ ", " + std::to_string(self.normal.z) + ")}"; }''')
+    @PYB11implementation('''[](const Plane<%(VectorType)s, %(VA)s>& self) { return "{" + std::to_string(self.dist) + ", (" + std::to_string(self.normal.x) + ", " + std::to_string(self.normal.y) + ")}"; }''')
     def __repr__(self):
         return
